requirements.txt

opencv-python-headless
gunicorn
itsdangerous

MarkupSafe
Werkzeug
numpy
scipy

matplotlib
pandas

keras
 h5py
Tensorflow-cpu 
streamlit
